fixed nits
